Move date parsing logic from ArchiveInvalidator to CoreAdminHome/API.php so the service class only accepts well formed input.
